--- a/crates/accelerate/src/target_transpiler/mod.rs
+++ b/crates/accelerate/src/target_transpiler/mod.rs
@@ -59,7 +59,7 @@ type GateMapState = Vec<(String, Vec<(Option<Qargs>, Option<InstructionPropertie
 /// Represents a Qiskit `Gate` object or a Variadic instruction.
 /// Keeps a reference to its Python instance for caching purposes.
 #[derive(FromPyObject, Debug, Clone, IntoPyObjectRef)]
-pub(crate) enum TargetOperation {
+pub enum TargetOperation {
     Normal(NormalOperation),
     Variadic(PyObject),
 }
@@ -89,7 +89,7 @@ impl TargetOperation {
 /// Represents a Qiskit `Gate` object, keeps a reference to its Python
 /// instance for caching purposes.
 #[derive(Debug, Clone)]
-pub(crate) struct NormalOperation {
+pub struct NormalOperation {
     pub operation: PackedOperation,
     pub params: SmallVec<[Param; 3]>,
     op_object: PyObject,
@@ -145,7 +145,7 @@ memory.
     module = "qiskit._accelerate.target"
 )]
 #[derive(Clone, Debug)]
-pub(crate) struct Target {
+pub struct Target {
     #[pyo3(get, set)]
     pub description: Option<String>,
     #[pyo3(get)]