--- a/crates/accelerate/src/target_transpiler/nullable_index_map.rs
+++ b/crates/accelerate/src/target_transpiler/nullable_index_map.rs
@@ -35,7 +35,7 @@ type BaseMap<K, V> = IndexMap<K, V, RandomState>;
 /// **Warning:** This is an experimental feature and should be used with care as it does not
 /// fully implement all the methods present in `IndexMap<K, V>` due to API limitations.
 #[derive(Debug, Clone, IntoPyObject, IntoPyObjectRef)]
-pub(crate) struct NullableIndexMap<K, V>
+pub struct NullableIndexMap<K, V>
 where
     K: Eq + Hash + Clone,
     V: Clone,