@mtreinish 2025-03-21
The one test that is failing: `test.python.compiler.test_transpiler.TestTranspile.test_scheduling_dt_constraints` looks to be hitting a bug elsewhere in Qiskit. The underlying issue is that the pass manager is constructed with: `transpile(qc, backend=backend_v2, scheduling_method="asap", dt=backend.target.dt / 2)` and the error rates from the target seem to disappear when vf2layout is passed the target object. This is causing the test to fail because it doesn't have a metric to pick a qubit from.

I think what is happening is that there is a `Target.from_configuration` call to override the instruction durations with a custom dt and that doesn't know about the error rates from the original target and it's simply omitting them.