@mtreinish 2025-03-21
The performance here is looking pretty good without any real tuning, especially considering most of the heavy lifting was in rust already,

```
Benchmarks that have improved:

| Change   | Before [e0da163e] <vf2-rs~3^2>   | After [daec870e] <vf2-rs>   |   Ratio | Benchmark (Parameter)                                   |
|----------|----------------------------------|-----------------------------|---------|---------------------------------------------------------|
| -        | 7.55±0.05ms                      | 6.32±0.06ms                 |    0.84 | utility_scale.UtilityScaleBenchmarks.time_bvlike('ecr') |
| -        | 7.61±0.02ms                      | 6.31±0.07ms                 |    0.83 | utility_scale.UtilityScaleBenchmarks.time_bvlike('cx')  |
| -        | 7.57±0.08ms                      | 6.32±0.07ms                 |    0.83 | utility_scale.UtilityScaleBenchmarks.time_bvlike('cz')  |

Benchmarks that have stayed the same:

| Change   | Before [e0da163e] <vf2-rs~3^2>   | After [daec870e] <vf2-rs>   | Ratio   | Benchmark (Parameter)                                                         |
|----------|----------------------------------|-----------------------------|---------|-------------------------------------------------------------------------------|
|          | 0                                | 0                           | n/a     | utility_scale.UtilityScaleBenchmarks.track_bvlike_depth('cx')                 |
|          | 0                                | 0                           | n/a     | utility_scale.UtilityScaleBenchmarks.track_bvlike_depth('cz')                 |
|          | 0                                | 0                           | n/a     | utility_scale.UtilityScaleBenchmarks.track_bvlike_depth('ecr')                |
|          | 94.2±0.9ms                       | 95.0±1ms                    | 1.01    | utility_scale.UtilityScaleBenchmarks.time_parse_qft_n100('cz')                |
|          | 30.4±0.2ms                       | 30.7±0.1ms                  | 1.01    | utility_scale.UtilityScaleBenchmarks.time_parse_square_heisenberg_n100('cz')  |
|          | 30.1±0.1ms                       | 30.5±0.2ms                  | 1.01    | utility_scale.UtilityScaleBenchmarks.time_parse_square_heisenberg_n100('ecr') |
|          | 8.78±0.06ms                      | 8.76±0.1ms                  | 1.00    | utility_scale.UtilityScaleBenchmarks.time_parse_qaoa_n100('cx')               |
|          | 8.80±0.04ms                      | 8.80±0.06ms                 | 1.00    | utility_scale.UtilityScaleBenchmarks.time_parse_qaoa_n100('cz')               |
|          | 95.1±0.5ms                       | 95.3±0.8ms                  | 1.00    | utility_scale.UtilityScaleBenchmarks.time_parse_qft_n100('ecr')               |
|          | 387                              | 387                         | 1.00    | utility_scale.UtilityScaleBenchmarks.track_bv_100_depth('cx')                 |
|          | 402                              | 402                         | 1.00    | utility_scale.UtilityScaleBenchmarks.track_bv_100_depth('cz')                 |
|          | 406                              | 406                         | 1.00    | utility_scale.UtilityScaleBenchmarks.track_bv_100_depth('ecr')                |
|          | 300                              | 300                         | 1.00    | utility_scale.UtilityScaleBenchmarks.track_circSU2_depth('cx')                |
|          | 300                              | 300                         | 1.00    | utility_scale.UtilityScaleBenchmarks.track_circSU2_depth('cz')                |
|          | 300                              | 300                         | 1.00    | utility_scale.UtilityScaleBenchmarks.track_circSU2_depth('ecr')               |
|          | 1727                             | 1727                        | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qaoa_depth('cx')                   |
|          | 1734                             | 1734                        | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qaoa_depth('cz')                   |
|          | 1734                             | 1734                        | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qaoa_depth('ecr')                  |
|          | 1982                             | 1982                        | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qft_depth('cx')                    |
|          | 1982                             | 1982                        | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qft_depth('cz')                    |
|          | 1980                             | 1980                        | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qft_depth('ecr')                   |
|          | 2694                             | 2694                        | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qv_depth('cx')                     |
|          | 2694                             | 2694                        | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qv_depth('cz')                     |
|          | 2694                             | 2694                        | 1.00    | utility_scale.UtilityScaleBenchmarks.track_qv_depth('ecr')                    |
|          | 558                              | 558                         | 1.00    | utility_scale.UtilityScaleBenchmarks.track_square_heisenberg_depth('cx')      |
|          | 558                              | 558                         | 1.00    | utility_scale.UtilityScaleBenchmarks.track_square_heisenberg_depth('cz')      |
|          | 558                              | 558                         | 1.00    | utility_scale.UtilityScaleBenchmarks.track_square_heisenberg_depth('ecr')     |
|          | 818±20ms                         | 812±20ms                    | 0.99    | utility_scale.UtilityScaleBenchmarks.time_circSU2('cz')                       |
|          | 826±20ms                         | 814±20ms                    | 0.99    | utility_scale.UtilityScaleBenchmarks.time_circSU2('ecr')                      |
|          | 95.1±0.4ms                       | 94.3±1ms                    | 0.99    | utility_scale.UtilityScaleBenchmarks.time_parse_qft_n100('cx')                |
|          | 30.9±0.2ms                       | 30.7±0.2ms                  | 0.99    | utility_scale.UtilityScaleBenchmarks.time_parse_square_heisenberg_n100('cx')  |
|          | 528±2ms                          | 523±2ms                     | 0.99    | utility_scale.UtilityScaleBenchmarks.time_qft('ecr')                          |
|          | 105±2ms                          | 104±0.7ms                   | 0.99    | utility_scale.UtilityScaleBenchmarks.time_square_heisenberg('ecr')            |
|          | 73.7±0.9ms                       | 71.8±1ms                    | 0.98    | utility_scale.UtilityScaleBenchmarks.time_bv_100('cz')                        |
|          | 73.5±0.6ms                       | 71.8±0.4ms                  | 0.98    | utility_scale.UtilityScaleBenchmarks.time_bv_100('ecr')                       |
|          | 9.00±0.06ms                      | 8.81±0.08ms                 | 0.98    | utility_scale.UtilityScaleBenchmarks.time_parse_qaoa_n100('ecr')              |
|          | 271±4ms                          | 265±1ms                     | 0.98    | utility_scale.UtilityScaleBenchmarks.time_qaoa('cz')                          |
|          | 408±2ms                          | 400±2ms                     | 0.98    | utility_scale.UtilityScaleBenchmarks.time_qft('cx')                           |
|          | 512±5ms                          | 500±3ms                     | 0.98    | utility_scale.UtilityScaleBenchmarks.time_qv('cx')                            |
|          | 619±4ms                          | 604±8ms                     | 0.98    | utility_scale.UtilityScaleBenchmarks.time_qv('cz')                            |
|          | 621±9ms                          | 606±2ms                     | 0.98    | utility_scale.UtilityScaleBenchmarks.time_qv('ecr')                           |
|          | 105±1ms                          | 103±1ms                     | 0.98    | utility_scale.UtilityScaleBenchmarks.time_square_heisenberg('cz')             |
|          | 821±20ms                         | 795±30ms                    | 0.97    | utility_scale.UtilityScaleBenchmarks.time_circSU2('cx')                       |
|          | 203±1ms                          | 197±3ms                     | 0.97    | utility_scale.UtilityScaleBenchmarks.time_qaoa('cx')                          |
|          | 257±1ms                          | 250±3ms                     | 0.97    | utility_scale.UtilityScaleBenchmarks.time_qaoa('ecr')                         |
|          | 70.9±0.7ms                       | 67.7±0.4ms                  | 0.96    | utility_scale.UtilityScaleBenchmarks.time_bv_100('cx')                        |
|          | 515±4ms                          | 495±5ms                     | 0.96    | utility_scale.UtilityScaleBenchmarks.time_qft('cz')                           |
|          | 92.0±1ms                         | 88.5±0.8ms                  | 0.96    | utility_scale.UtilityScaleBenchmarks.time_square_heisenberg('cx')             |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```